move most dump changes to rapids conf
- fork generation step with maven.compile.classpath
- change to a phase before package
