+++ integration_tests/src/main/scala/com/nvidia/spark/rapids/tests/DumpDefaultConfigs.scala
@@ -16,66 +16,21 @@
 
 package com.nvidia.spark.rapids.tests
 
-import java.io.{BufferedOutputStream, DataOutputStream, FileOutputStream}
-import java.nio.charset.StandardCharsets
-import java.util.Locale
+import org.apache.commons.lang3.reflect.MethodUtils
 
-import com.nvidia.spark.rapids.Arm.withResource
 import com.nvidia.spark.rapids.ShimReflectionUtils
-import org.json4s.DefaultFormats
-import org.json4s.jackson.Serialization.writePretty
 
 /**
  * Dump all spark-rapids configs with their defaults.
  */
 object DumpDefaultConfigs {
-
-  object Format extends Enumeration {
-    type Format = Value
-    val PLAIN, JSON = Value
-  }
-
   def main(args: Array[String]): Unit = {
     if (args.length < 2) {
       System.err.println(s"Usage: ${this.getClass.getCanonicalName} {format} {output_path}")
       System.exit(1)
     }
-
-    val format = Format.withName(args(0).toUpperCase(Locale.US))
-    val outputPath = args(1)
-
-    println(s"Dumping all spark-rapids configs and their defaults at ${outputPath}")
-
     // We use the reflection as RapidsConf should be accessed via the shim layer.
     val clazz = ShimReflectionUtils.loadClass("com.nvidia.spark.rapids.RapidsConf")
-    val m = clazz.getDeclaredMethod("getAllConfigsWithDefault")
-    val allConfs: Map[String, Any] = m.invoke(null).asInstanceOf[Map[String, Any]]
-    withResource(new FileOutputStream(outputPath)) { fos =>
-      withResource(new BufferedOutputStream(fos)) { bos =>
-        format match {
-          case Format.PLAIN =>
-            withResource(new DataOutputStream(bos)) { dos =>
-              allConfs.foreach( { case (k, v) =>
-                val valStr = v match {
-                  case Some(optVal) => optVal.toString
-                  case None => ""
-                  case _ =>
-                    if (v == null) {
-                      ""
-                    } else {
-                      v.toString
-                    }
-                }
-                dos.writeUTF(s"'${k}': '${valStr}',")
-              })
-            }
-          case Format.JSON =>
-            implicit val formats = DefaultFormats
-            bos.write(writePretty(allConfs).getBytes(StandardCharsets.UTF_8))
-          case _ =>
-            System.err.println(s"Unknown format: ${format}")
-        }
-      }
-    }
+    MethodUtils.invokeStaticMethod(clazz, "dumpConfigsWithDefault", args)
   }
 }
